[BL] Move propertyTreeToString into ConfigTree.
